Add scripts for opening a Wikipedia link, starting HTTP server and tools for Lab 6
Two new Python scripts were added. The first is designed for opening a specific Wikipedia link in the browser, finding an available starting point port from 8000, and starting a HTTP server on that port. The second script includes a user interface with Tkinter for starting and stopping the server, as well as opening the extension URL and the server URL.

These changes offer a quicker setup for running a simple HTTP server and install the CheerpJ plugin. The plugin is needed to run the simulation software which is used in Lab 6. Additionally, various files were added to "src/Lab_6" to cover the lab's content.